backend/requirements.txt
pdfplumber
arabic-reshaper
python-bidi
langdetect
mistralai
flask
flask-cors
